requirements.txt
gradio==4.12.0
phidata==2.3.5
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
google-generativeai==0.3.2